+++ src/website/src/pages/ec/content-types/funding-programme/docs/code.mdx
@@ -4,7 +4,7 @@ order: 2
 ---
 
 import Playground from '@ecl/website-components/Playground';
-import FundingProgrammePageExample from '@ecl/ec-react-funding-programme-page/examples/Default';
+import FundingProgrammePageExample from '@ecl/ec-react-funding-programme-page/examples/FundingProgrammeCore';
 
 <Playground
   system="ec"
